Moved the camera to a camera subfolder with the manager
