+++ Weave_Engine/Weave_Engine/Weave_Engine.vcxproj.filters
@@ -58,6 +58,9 @@
     <Filter Include="Weave_Common\Utils">
       <UniqueIdentifier>{6c55ad0f-5c7f-45e9-bebb-44f222d9c4d0}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Camera">
+      <UniqueIdentifier>{3744b0aa-1c80-4f98-80e0-b382b43fdee5}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Core\DXCore.cpp">
@@ -68,9 +71,6 @@
     </ClCompile>
     <ClCompile Include="Core\Main.cpp">
       <Filter>Core</Filter>
-    </ClCompile>
-    <ClCompile Include="Entity\Camera.cpp">
-      <Filter>Entity</Filter>
     </ClCompile>
     <ClCompile Include="Entity\Entity.cpp">
       <Filter>Entity</Filter>
@@ -153,9 +153,6 @@
     <ClCompile Include="Scenes\Scene.cpp">
       <Filter>Scenes</Filter>
     </ClCompile>
-    <ClCompile Include="Core\CameraManager.cpp">
-      <Filter>Core</Filter>
-    </ClCompile>
     <ClCompile Include="Scripting\ScriptComponent.cpp">
       <Filter>Scripting</Filter>
     </ClCompile>
@@ -174,6 +171,12 @@
     <ClCompile Include="..\..\Weave_Common\Timing.cpp">
       <Filter>Weave_Common</Filter>
     </ClCompile>
+    <ClCompile Include="Camera\Camera.cpp">
+      <Filter>Camera</Filter>
+    </ClCompile>
+    <ClCompile Include="Camera\CameraManager.cpp">
+      <Filter>Camera</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Core\DXCore.h">
@@ -181,9 +184,6 @@
     </ClInclude>
     <ClInclude Include="Core\Game.h">
       <Filter>Core</Filter>
-    </ClInclude>
-    <ClInclude Include="Entity\Camera.h">
-      <Filter>Entity</Filter>
     </ClInclude>
     <ClInclude Include="Entity\Entity.h">
       <Filter>Entity</Filter>
@@ -293,9 +293,6 @@
     <ClInclude Include="Scenes\Scene.h">
       <Filter>Scenes</Filter>
     </ClInclude>
-    <ClInclude Include="Core\CameraManager.h">
-      <Filter>Core</Filter>
-    </ClInclude>
     <ClInclude Include="Scripting\ScriptComponent.h">
       <Filter>Scripting</Filter>
     </ClInclude>
@@ -331,6 +328,12 @@
     </ClInclude>
     <ClInclude Include="..\..\Weave_Common\Utils\Singleton.hpp">
       <Filter>Weave_Common\Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="Camera\Camera.h">
+      <Filter>Camera</Filter>
+    </ClInclude>
+    <ClInclude Include="Camera\CameraManager.h">
+      <Filter>Camera</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
